requirements.txt
tqdm
nltk
numpy
scikit-learn
msgpack-python
tensorboardX
json5
